Refactor ui slice to use feature folders
